requirements.txt
whoosh
dulwich
markdown
wheezy.web[wheezy.template]
jsonpatch
pyuwsgi
pdfminer.six
pymdown-extensions
Pygments
odfdo
ruamel.yaml
olefile
